+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/form_builder/form_builder_dd_support.js
@@ -208,6 +208,10 @@ AUI.add(
 
 						instance._addToStack(fieldNodeStart);
 
+						if (instance._sidebar.isOpen()) {
+							instance._sidebar.close();
+						}
+
 						if (proxyActive) {
 							proxyActive.empty();
 							proxyActive.append(clonedNode);
@@ -374,13 +378,15 @@ AUI.add(
 				var instance = this;
 
 				if (activeLayout.get('rows').length == 1 && instance._verifyEmptyForm(activeLayout.get('rows')[0]) && !A.one('#lfr-initial-col-message')) {
+					var columnNode = activeLayout.get('rows')[0].get('cols')[0].get('node');
 					var columnMessageNode = A.Node.create('<div/>');
 
 					columnMessageNode.text(Liferay.Language.get('drag-from-sidebar-and-drop-here'));
 					columnMessageNode.setAttribute('id', 'lfr-initial-col-message');
 
-					activeLayout.get('rows')[0].get('cols')[0].get('node').addClass('lfr-initial-col');
-					activeLayout.get('rows')[0].get('cols')[0].get('node').append(columnMessageNode);
+					columnNode.get('parentNode').all('.layout-builder-add-col-draggable').remove();
+					columnNode.addClass('lfr-initial-col');
+					columnNode.append(columnMessageNode);
 				}
 			},