+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -3089,9 +3089,14 @@ marker file.
 ### Changed
 - [LPS-78741]: Configure the `npmInstall` task to run `npm install` instead of
 `npm ci`.
-- [LPS-78741]: Update the [Liferay Gradle Plugins] dependency to version 3.9.8.
+- [LPS-78741]: Update the [Liferay Gradle Plugins] dependency to version 3.9.10.
 - [LPS-78741]: Update the [Liferay Gradle Plugins JSDoc] dependency to version
 2.0.3.
+
+## 5.3.44 - 2018-03-23
+
+### Changed
+- [LPS-78911]: Update the [Liferay Gradle Plugins] dependency to version 3.9.11.
 
 [Find Security Bugs]: https://github.com/liferay/liferay-portal/tree/master/modules/third-party/com-h3xstream-findsecbugs
 [Gradle License Report]: https://github.com/jk1/Gradle-License-Report